Regiinal parameters config added

+++ device-history-service-json.cpp
@@ -233,7 +233,7 @@ void JsonFileDeviceHistoryService::list(
 ) {
 	int64_t c = -1;
 	if (size == 0)
-		size = UINT64_MAX;
+		size = SIZE_MAX;
 	for (std::map<DEVADDRINT, DEVICE_HISTORY_ITEM>::const_iterator it(storage.begin()); it != storage.end(); it++) {
 		c++;
 		if (c < offset)
