+++ identity-service-file-json.cpp
@@ -315,7 +315,7 @@ void JsonFileIdentityService::list(
 ) {
     int64_t c = -1;
     if (size == 0)
-        size = UINT64_MAX;
+        size = SIZE_MAX;
     for (std::map<DEVADDRINT, DEVICEID>::const_iterator it(storage.begin()); it != storage.end(); it++) {
         c++;
         if (c < offset)
